refactor: first phase of generalizing installers
