+++ pkg/fab/recipe/control_build.go
@@ -8,16 +8,13 @@ import (
 	"crypto/sha256"
 	_ "embed"
 	"encoding/base64"
-	"errors"
 	"fmt"
 	"log/slog"
 	"os"
 	"path/filepath"
-	"slices"
 
 	fabapi "go.githedgehog.com/fabricator/api/fabricator/v1beta1"
 	"go.githedgehog.com/fabricator/pkg/artificer"
-	"go.githedgehog.com/fabricator/pkg/embed/recipebin"
 	"go.githedgehog.com/fabricator/pkg/fab/comp"
 	"go.githedgehog.com/fabricator/pkg/fab/comp/certmanager"
 	"go.githedgehog.com/fabricator/pkg/fab/comp/f8r"
@@ -38,16 +35,6 @@ import (
 //go:embed control_butane.tmpl.yaml
 var controlButaneTmpl string
 
-type BuildMode string
-
-const (
-	BuildModeManual BuildMode = "manual"
-	BuildModeUSB    BuildMode = "usb"
-	BuildModeISO    BuildMode = "iso"
-)
-
-var BuildModes = []BuildMode{BuildModeManual, BuildModeUSB, BuildModeISO}
-
 type ControlInstallBuilder struct {
 	WorkDir    string
 	Fab        fabapi.Fabricator
@@ -59,16 +46,9 @@ type ControlInstallBuilder struct {
 }
 
 const (
-	FabName                      = "fab.yaml"
-	WiringName                   = "wiring.yaml"
-	InstallSuffix                = "-install"
-	InstallArchiveSuffix         = InstallSuffix + ".tgz"
-	InstallIgnitionSuffix        = InstallSuffix + ".ign"
-	InstallUSBImageWorkdirSuffix = InstallSuffix + "-usb.wip"
-	InstallUSBImageSuffix        = InstallSuffix + "-usb.img"
-	InstallISOImageSuffix        = InstallSuffix + "-usb.iso"
-	InstallHashSuffix            = InstallSuffix + ".inhash"
-	RecipeBin                    = "hhfab-recipe"
+	BuildTypeControl = "control"
+	FabName          = "fab.yaml"
+	WiringName       = "wiring.yaml"
 )
 
 var AirgapArtifactLists = []comp.ListOCIArtifacts{
@@ -82,76 +62,26 @@ var AirgapArtifactLists = []comp.ListOCIArtifacts{
 }
 
 func (b *ControlInstallBuilder) Build(ctx context.Context) error {
-	if !slices.Contains(BuildModes, b.Mode) {
-		return fmt.Errorf("invalid build mode %q", b.Mode) //nolint:goerr113
-	}
-
-	installDir := filepath.Join(b.WorkDir, b.Control.Name+InstallSuffix)
-	installArchive := filepath.Join(b.WorkDir, b.Control.Name+InstallArchiveSuffix)
-	installIgnition := filepath.Join(b.WorkDir, b.Control.Name+InstallIgnitionSuffix)
-	installHashFile := filepath.Join(b.WorkDir, b.Control.Name+InstallHashSuffix)
-
-	newHash, err := b.hash(ctx)
+	hash, err := b.hash(ctx)
 	if err != nil {
-		return fmt.Errorf("hashing: %w", err)
+		return fmt.Errorf("hashing build config: %w", err)
 	}
 
-	if existingHash, err := os.ReadFile(installHashFile); err == nil {
-		slog.Debug("Checking existing installers", "new", newHash, "existing", string(existingHash))
+	return buildInstall(ctx, buildInstallOpts{
+		WorkDir:               b.WorkDir,
+		Name:                  b.Control.Name,
+		Type:                  BuildTypeControl,
+		Mode:                  b.Mode,
+		Hash:                  hash,
+		AddPayload:            b.addPayload,
+		BuildIgnition:         b.controlIgnition,
+		Downloader:            b.Downloader,
+		FlatcarUSBRootVersion: b.Fab.Status.Versions.Fabricator.ControlUSBRoot,
+	})
+}
 
-		files := []string{installDir, installArchive, installIgnition}
-		if b.Mode == BuildModeUSB {
-			files = []string{installDir, filepath.Join(b.WorkDir, b.Control.Name+InstallUSBImageSuffix)}
-		}
-		if b.Mode == BuildModeISO {
-			files = []string{installDir, filepath.Join(b.WorkDir, b.Control.Name+InstallISOImageSuffix)}
-		}
-		if string(existingHash) == newHash && isPresent(files...) {
-			slog.Info("Using existing installers")
-
-			return nil
-		}
-	}
-
-	if err := removeIfExists(installHashFile); err != nil {
-		return fmt.Errorf("removing hash file: %w", err)
-	}
-
-	slog.Info("Building installer", "control", b.Control.Name)
-
-	if err := removeIfExists(installDir); err != nil {
-		return fmt.Errorf("removing install dir: %w", err)
-	}
-	if err := removeIfExists(installArchive); err != nil {
-		return fmt.Errorf("removing install archive: %w", err)
-	}
-	if err := removeIfExists(installIgnition); err != nil {
-		return fmt.Errorf("removing install ignition: %w", err)
-	}
-	if err := removeIfExists(filepath.Join(b.WorkDir, b.Control.Name+InstallUSBImageWorkdirSuffix)); err != nil {
-		return fmt.Errorf("removing install usb image workdir: %w", err)
-	}
-	if err := removeIfExists(filepath.Join(b.WorkDir, b.Control.Name+InstallUSBImageSuffix)); err != nil {
-		return fmt.Errorf("removing install usb image: %w", err)
-	}
-	if err := removeIfExists(filepath.Join(b.WorkDir, b.Control.Name+InstallISOImageSuffix)); err != nil {
-		return fmt.Errorf("removing install iso image: %w", err)
-	}
-
-	if err := os.MkdirAll(installDir, 0o700); err != nil {
-		return fmt.Errorf("creating install dir: %w", err)
-	}
-
-	slog.Info("Adding recipe bin to installer", "control", b.Control.Name)
-	recipeBin, err := recipebin.Bytes()
-	if err != nil {
-		return fmt.Errorf("getting recipe bin: %w", err)
-	}
-	if err := os.WriteFile(filepath.Join(installDir, RecipeBin), recipeBin, 0o700); err != nil { //nolint:gosec
-		return fmt.Errorf("writing recipe bin: %w", err)
-	}
-
-	slog.Info("Adding k3s and tools to installer", "control", b.Control.Name)
+func (b *ControlInstallBuilder) addPayload(ctx context.Context, slog *slog.Logger, installDir string) error {
+	slog.Info("Adding k3s and tools to installer")
 	if err := b.Downloader.FromORAS(ctx, installDir, k3s.Ref, k3s.Version(b.Fab), []artificer.ORASFile{
 		{
 			Name: k3s.BinName,
@@ -175,7 +105,7 @@ func (b *ControlInstallBuilder) Build(ctx context.Context) error {
 		return fmt.Errorf("downloading k9s: %w", err)
 	}
 
-	slog.Info("Adding zot to installer", "control", b.Control.Name)
+	slog.Info("Adding zot to installer")
 	if err := b.Downloader.FromORAS(ctx, installDir, zot.AirgapRef, zot.Version(b.Fab), []artificer.ORASFile{
 		{
 			Name: zot.AirgapImageName,
@@ -187,7 +117,7 @@ func (b *ControlInstallBuilder) Build(ctx context.Context) error {
 		return fmt.Errorf("downloading zot: %w", err)
 	}
 
-	slog.Info("Adding Flatcar Upgrade bin to installer", "control", b.Control.Name)
+	slog.Info("Adding flatcar upgrade bin to installer")
 	if err := b.Downloader.FromORAS(ctx, installDir, flatcar.UpdateRef, flatcar.Version(b.Fab), []artificer.ORASFile{
 		{
 			Name: flatcar.UpdateBinName,
@@ -196,7 +126,7 @@ func (b *ControlInstallBuilder) Build(ctx context.Context) error {
 		return fmt.Errorf("downloading flatcar-update: %w", err)
 	}
 
-	slog.Info("Adding cert-manager to installer", "control", b.Control.Name)
+	slog.Info("Adding cert-manager to installer")
 	if err := b.Downloader.FromORAS(ctx, installDir, certmanager.AirgapRef, certmanager.Version(b.Fab), []artificer.ORASFile{
 		{
 			Name: certmanager.AirgapImageName,
@@ -208,7 +138,7 @@ func (b *ControlInstallBuilder) Build(ctx context.Context) error {
 		return fmt.Errorf("downloading cert-manager: %w", err)
 	}
 
-	slog.Info("Adding config and wiring files to installer", "control", b.Control.Name)
+	slog.Info("Adding config and wiring files to installer")
 	fabF, err := os.Create(filepath.Join(installDir, FabName))
 	if err != nil {
 		return fmt.Errorf("creating fab file: %w", err)
@@ -248,7 +178,7 @@ func (b *ControlInstallBuilder) Build(ctx context.Context) error {
 	}
 
 	if b.Fab.Spec.Config.Registry.IsAirgap() {
-		slog.Info("Adding airgap artifacts to installer", "control", b.Control.Name)
+		slog.Info("Adding airgap artifacts to installer")
 
 		airgapArts, err := comp.CollectArtifacts(b.Fab, AirgapArtifactLists...)
 		if err != nil {
@@ -262,52 +192,16 @@ func (b *ControlInstallBuilder) Build(ctx context.Context) error {
 		}
 	}
 
-	if b.Mode == BuildModeUSB || b.Mode == BuildModeISO {
-		if err := b.buildUSBImage(ctx); err != nil {
-			return fmt.Errorf("building USB image: %w", err)
-		}
-	} else {
-		slog.Info("Archiving installer", "path", installArchive, "control", b.Control.Name)
-		if err := archiveTarGz(ctx, installDir, installArchive); err != nil {
-			return fmt.Errorf("archiving install: %w", err)
-		}
-
-		slog.Info("Creating ignition", "path", installIgnition, "control", b.Control.Name)
-		ign, err := controlIgnition(b.Fab, b.Control, "")
-		if err != nil {
-			return fmt.Errorf("creating ignition: %w", err)
-		}
-
-		if err := os.WriteFile(installIgnition, ign, 0o600); err != nil {
-			return fmt.Errorf("writing ignition: %w", err)
-		}
-	}
-
-	if err := os.WriteFile(installHashFile, []byte(newHash), 0o600); err != nil {
-		return fmt.Errorf("writing hash: %w", err)
-	}
-
 	return nil
 }
 
-func removeIfExists(path string) error {
-	_, err := os.Stat(path)
-	if err != nil && errors.Is(err, os.ErrNotExist) {
-		return nil
-	}
-	if err != nil {
-		return fmt.Errorf("checking %q: %w", path, err)
+func (b *ControlInstallBuilder) controlIgnition() ([]byte, error) {
+	autoInstallPath := filepath.Join(OSTargetInstallDir, BuildTypeControl+Separator+b.Control.Name+Separator+InstallSuffix)
+	if b.Mode == BuildModeManual {
+		autoInstallPath = ""
 	}
 
-	if err := os.RemoveAll(path); err != nil {
-		return fmt.Errorf("removing %q: %w", path, err)
-	}
-
-	return nil
-}
-
-func controlIgnition(fab fabapi.Fabricator, control fabapi.ControlNode, autoInstall string) ([]byte, error) {
-	dummyIP, err := control.Spec.Dummy.IP.Parse()
+	dummyIP, err := b.Control.Spec.Dummy.IP.Parse()
 	if err != nil {
 		return nil, fmt.Errorf("parsing dummy IP: %w", err)
 	}
@@ -316,19 +210,19 @@ func controlIgnition(fab fabapi.Fabricator, control fabapi.ControlNode, autoInst
 	}
 
 	but, err := tmplutil.FromTemplate("butane", controlButaneTmpl, map[string]any{
-		"Hostname":       control.Name,
-		"PasswordHash":   fab.Spec.Config.Control.DefaultUser.PasswordHash,
-		"AuthorizedKeys": fab.Spec.Config.Control.DefaultUser.AuthorizedKeys,
-		"MgmtInterface":  control.Spec.Management.Interface,
-		"MgmtAddress":    control.Spec.Management.IP,
-		"ControlVIP":     fab.Spec.Config.Control.VIP,
-		"ExtInterface":   control.Spec.External.Interface,
-		"ExtAddress":     control.Spec.External.IP,
-		"ExtGateway":     control.Spec.External.Gateway,
-		"ExtDNS":         control.Spec.External.DNS,
+		"Hostname":       b.Control.Name,
+		"PasswordHash":   b.Fab.Spec.Config.Control.DefaultUser.PasswordHash,
+		"AuthorizedKeys": b.Fab.Spec.Config.Control.DefaultUser.AuthorizedKeys,
+		"MgmtInterface":  b.Control.Spec.Management.Interface,
+		"MgmtAddress":    b.Control.Spec.Management.IP,
+		"ControlVIP":     b.Fab.Spec.Config.Control.VIP,
+		"ExtInterface":   b.Control.Spec.External.Interface,
+		"ExtAddress":     b.Control.Spec.External.IP,
+		"ExtGateway":     b.Control.Spec.External.Gateway,
+		"ExtDNS":         b.Control.Spec.External.DNS,
 		"DummyAddress":   dummyIP.Masked().String(),
 		"DummyGateway":   dummyIP.Masked().Addr().Next().String(),
-		"AutoInstall":    autoInstall,
+		"AutoInstall":    autoInstallPath,
 	})
 	if err != nil {
 		return nil, fmt.Errorf("butane: %w", err)
@@ -362,14 +256,4 @@ func (b *ControlInstallBuilder) hash(ctx context.Context) (string, error) {
 	}
 
 	return base64.URLEncoding.EncodeToString(h.Sum(nil)), nil
-}
-
-func isPresent(files ...string) bool {
-	for _, f := range files {
-		if _, err := os.Stat(f); err != nil {
-			return false
-		}
-	}
-
-	return true
 }
