+++ pkg/fab/recipe/control_build_usb_butane.yaml
@@ -38,4 +38,3 @@ systemd:
         [Install]
         WantedBy=local-fs.target
         RequiredBy=flatcar-install.service
-
